services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "ycalk",email = "alosev752@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "faststream[rabbit] (>=0.5.47,<0.6.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "ai-image-detector-messaging-schema (>=0.1.0,<0.2.0)"
]

[tool.poetry]
packages = [{include = "api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
api = "api:run"